Splits menu-dropdown out into its own file so that the site's doc gen script can read it.
